third_party/blink/web_tests/http/tests/credentialmanager/credentialscontainer-create-with-virtual-authenticator.html

<!DOCTYPE html>
<title>Credential Manager: End-to-end tests for create() with a virtual authenticator.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings_lite.js"></script>
<script src="/gen/mojo/public/mojom/base/time.mojom-lite.js"></script>
<script src="/gen/url/mojom/url.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/authenticator.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/virtual_authenticator.mojom-lite.js"></script>
<script src="resources/test-inputs.js"></script>
<script src="resources/virtual-navigator-credentials.js"></script>
<body>
<script>

if (document.location.host != "subdomain.example.test:8443") {
  document.location = "https://subdomain.example.test:8443/credentialmanager/credentialscontainer-create-with-virtual-authenticator.html";
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
}

promise_test(async _ => {
  let authenticators = await navigator.credentials.test.authenticators();
  assert_equals(authenticators.length, 0);
  let testAuthenticator = await navigator.credentials.test.createAuthenticator();
}, "Set up the testing environment.");

promise_test(async t => {
  let authenticators = await navigator.credentials.test.authenticators();
  assert_equals(authenticators.length, 1);
  let testAuthenticator = authenticators[0];
  assert_true(await testAuthenticator.generateAndRegisterKey(ACCEPTABLE_CREDENTIAL_ID, "subdomain.example.test"));

  var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
  customMakeCredOptions.excludeCredentials = [ACCEPTABLE_CREDENTIAL];
  return promise_rejects(t, "InvalidStateError",
    navigator.credentials.create({ publicKey : customMakeCredOptions}));
}, "navigator.credentials.create() re-registration returns InvalidStateError");

promise_test(async t => {
  var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
  customMakeCredOptions.rp.id = "";
  return promise_rejects(t, "SecurityError",
    navigator.credentials.create({ publicKey : customMakeCredOptions}));
}, "navigator.credentials.create() with empty rpId returns SecurityError");

promise_test(async t => {
  var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
  customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
  return navigator.credentials.create({ publicKey : customMakeCredOptions });
}, "navigator.credentials.create() with resident keys is supported");

promise_test(async t => {
  authAbortController = new AbortController();
  authAbortSignal = authAbortController.signal;
  return navigator.credentials.create({ publicKey : MAKE_CREDENTIAL_OPTIONS, signal : authAbortSignal});
}, "navigator.credentials.create() with abort signal flag not set");

promise_test(async t => {
  authAbortController = new AbortController();
  authAbortSignal = authAbortController.signal;
  authAbortController.abort();
  var promise = navigator.credentials.create({ publicKey : MAKE_CREDENTIAL_OPTIONS, signal : authAbortSignal});
  return promise_rejects(t, "AbortError", promise);
}, "navigator.credentials.create() with abort signal flag set before sending request");

promise_test(async t => {
  authAbortController = new AbortController();
  authAbortSignal = authAbortController.signal;
  var promise = navigator.credentials.create({ publicKey : MAKE_CREDENTIAL_OPTIONS, signal : authAbortSignal});
  authAbortController.abort();
  return promise_rejects(t, "AbortError", promise);
}, "navigator.credentials.create() with abort signal flag is set after sending request");

promise_test(async t => {
  authAbortController = new AbortController();
  authAbortSignal = authAbortController.signal;
  let promise = await navigator.credentials.create({ publicKey : MAKE_CREDENTIAL_OPTIONS, signal : authAbortSignal});
  authAbortController.abort();
  return promise;
}, "navigator.credentials.create() with abort signal flag is set after promise resolved");

promise_test(t => {
  return navigator.credentials.test.clearAuthenticators();
}, "Clean up testing environment.");

</script>
